pyproject.toml
[project]
name = "project-02-09-26"
version = "0.1.0"
description = "Financial Data Analytics Dashboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair<5",
    "certifi>=2026.1.4",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pydotenv>=0.0.7",
    "python-certifi-win32>=1.6.1",
    "streamlit>=1.19.0",
    "urllib3>=1.26.18",
    "yfinance>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "python-dotenv>=1.2.1"
]
